+++ __tests__/util/codemirror-utils.test.js
@@ -20,7 +20,10 @@ const codeMirror = {
 };
 
 const mockAllRules = {
-  rules: {},
+  rules: {
+    'str': {prettyName: "String", color: "#FFFFFF"},
+    'number': {prettyName: "Number", color: "#FFFFFF"}
+  },
   ignoredRules: []
 }
 
@@ -100,6 +103,63 @@ describe('util: codemirror-utils', () => {
       const parser = 'python3';
       const expected = 'python';
       expect(codemirrorUtils.getCodeMirrorMode(parser)).toEqual(expected);
+    });
+  });
+  describe('generateFilters()', () => {
+    it('should return {} is ruleCounts is undefined', () => {
+      const prevFilters = {};
+      const ruleCounts = undefined;
+      expect(codemirrorUtils.generateFilters(prevFilters, ruleCounts)).toEqual({});
+    });
+    it('should return {} is ruleCounts is empty', () => {
+      const prevFilters = {};
+      const ruleCounts = {};
+      expect(codemirrorUtils.generateFilters(prevFilters, ruleCounts)).toEqual({});
+    });
+    it('does not contain rules that are not in the rules data object', () => {
+      const prevFilters = {};
+      const ruleCounts = {
+        'Rick Sanchez': 1,
+        'Morty Smith': 2,
+        'Summer Smith': 3,
+        'Beth Smith': 4,
+        'Jerry Smith': 5,
+      };
+      expect(codemirrorUtils.generateFilters(prevFilters, ruleCounts)).toEqual({});
+    });
+    it('sets selected property to false if filter not in prevFilters', () => {
+      const prevFilters = {};
+      const ruleCounts = {
+        str: 1,
+      };
+      const newFilters = codemirrorUtils.generateFilters(prevFilters, ruleCounts);
+      expect(newFilters.str.selected).toEqual(false);
+    });
+    it('sets selected property to filter\'s value in prevFilters', () => {
+      const prevFilters = {
+        str: { selected: true, },
+        number: { selected: false, },
+      };
+      const ruleCounts = {
+        str: 1,
+        number: 2,
+      };
+      const newFilters = codemirrorUtils.generateFilters(prevFilters, ruleCounts);
+      expect(newFilters.str.selected).toEqual(true);
+      expect(newFilters.number.selected).toEqual(false);
+    });
+    it('sets the filter\'s count property from ruleCounts', () => {
+      const prevFilters = {
+        str: { selected: true, },
+        number: { selected: false, },
+      };
+      const ruleCounts = {
+        str: 1,
+        number: 2,
+      };
+      const newFilters = codemirrorUtils.generateFilters(prevFilters, ruleCounts);
+      expect(newFilters.str.count).toEqual(1);
+      expect(newFilters.number.count).toEqual(2);
     });
   });
 });
