+++ __tests__/util/rules.test.js
@@ -60,61 +60,4 @@ describe('util: rules', () => {
       expect(ruleCounts[nodeType]).toEqual(3);
     });
   });
-  describe('generateFilters()', () => {
-    it('should return {} is ruleCounts is undefined', () => {
-      const prevFilters = {};
-      const ruleCounts = undefined;
-      expect(rules.generateFilters(prevFilters, ruleCounts)).toEqual({});
-    });
-    it('should return {} is ruleCounts is empty', () => {
-      const prevFilters = {};
-      const ruleCounts = {};
-      expect(rules.generateFilters(prevFilters, ruleCounts)).toEqual({});
-    });
-    it('does not contain rules that are not in the rules data object', () => {
-      const prevFilters = {};
-      const ruleCounts = {
-        'Rick Sanchez': 1,
-        'Morty Smith': 2,
-        'Summer Smith': 3,
-        'Beth Smith': 4,
-        'Jerry Smith': 5,
-      };
-      expect(rules.generateFilters(prevFilters, ruleCounts)).toEqual({});
-    });
-    it('sets selected property to false if filter not in prevFilters', () => {
-      const prevFilters = {};
-      const ruleCounts = {
-        str: 1,
-      };
-      const newFilters = rules.generateFilters(prevFilters, ruleCounts);
-      expect(newFilters.str.selected).toEqual(false);
-    });
-    it('sets selected property to filter\'s value in prevFilters', () => {
-      const prevFilters = {
-        str: { selected: true, },
-        number: { selected: false, },
-      };
-      const ruleCounts = {
-        str: 1,
-        number: 2,
-      };
-      const newFilters = rules.generateFilters(prevFilters, ruleCounts);
-      expect(newFilters.str.selected).toEqual(true);
-      expect(newFilters.number.selected).toEqual(false);
-    });
-    it('sets the filter\'s count property from ruleCounts', () => {
-      const prevFilters = {
-        str: { selected: true, },
-        number: { selected: false, },
-      };
-      const ruleCounts = {
-        str: 1,
-        number: 2,
-      };
-      const newFilters = rules.generateFilters(prevFilters, ruleCounts);
-      expect(newFilters.str.count).toEqual(1);
-      expect(newFilters.number.count).toEqual(2);
-    });
-  });
 });
